% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Calculates the shortest path from the initial node to every other node in the graph. 
For further information about the Euclidean Algorithm see here: https://en.wikipedia.org/wiki/Dijkstra%27s}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A graph that has.}

\item{init_node}{number.}
}
\value{
The greatest common divisor
}
\description{
Calculates the shortest path from the initial node to every other node in the graph. 
For further information about the Euclidean Algorithm see here: https://en.wikipedia.org/wiki/Dijkstra%27s
}
\examples{
wiki_graph <-
data.frame(v1=c(1,1,1,2,2,2,3,3,3,3,4,4,4,5,5,6,6,6),
v2=c(2,3,6,1,3,4,1,2,4,6,2,3,5,4,6,1,3,5),
w=c(7,9,14,7,10,15,9,10,11,2,15,11,6,6,9,14,2,9))
dijkstra(wiki_graph, 1)
dijkstra(wiki_graph, 3)
}
