% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean algorithm to find the greatest common divisor of two numbers.}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{is a numeric scalar or integer.}

\item{b}{is a numeric scalar or integer.}
}
\value{
Will return the greatest common divisor between a and b
}
\description{
This function takes two numeric scalars or integers and find the greatest
common divisor between them.
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)
}
\references{
For further information about the Euclidean algorithm see here: 
\url{https://en.wikipedia.org/wiki/Euclidean}
}
