% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab3Euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Find the greatest common divisor of two numbers}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A number.}

\item{b}{A number.}
}
\value{
The greatest common divisor
}
\description{
Find the greatest common divisor of two numbers
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)
}
