% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Find the greatest common divisor of two numbers. 
For further information about the Euclidean Algorithm see here: https://en.wikipedia.org/wiki/Euclidean}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A number.}

\item{b}{A number.}
}
\value{
The greatest common divisor
}
\description{
Find the greatest common divisor of two numbers. 
For further information about the Euclidean Algorithm see here: https://en.wikipedia.org/wiki/Euclidean
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)
}
